image: setting for rk356x, extlinux.conf is used as boot configuration

+++ build_images.sh
@@ -358,6 +358,16 @@ create_initrd() {
 setting_bootloader() {
     message "" "setting" "bootloader"
     # u-boot config
+    if [[ -f $CWD/config/boot_scripts/extlinux-$SOCFAMILY.conf ]]; then
+        install -Dm644 $CWD/config/boot_scripts/extlinux-$SOCFAMILY.conf "$SOURCE/$ROOTFS/boot/extlinux.conf"
+        # u-boot serial inteface config
+        sed -e "s:%ROOT_DISK%:${ROOT_DISK}:g" \
+            -e "s:%DEVICE_TREE_BLOB%:${DEVICE_TREE_BLOB}:g" \
+            -e "s:%SERIAL_CONSOLE%:${SERIAL_CONSOLE}:g" \
+            -e "s:%SERIAL_CONSOLE_SPEED%:${SERIAL_CONSOLE_SPEED}:g" \
+            -i "$SOURCE/$ROOTFS/boot/extlinux.conf"
+        return 0
+    fi
     if [[ -f $CWD/config/boot_scripts/boot-$SOCFAMILY.cmd ]]; then
         install -Dm644 $CWD/config/boot_scripts/boot-$SOCFAMILY.cmd "$SOURCE/$ROOTFS/boot/boot.cmd"
         # u-boot serial inteface config
